make MusicLibraryTrack strictly a data object
