+++ MusicDatabaseGenerator.csproj
@@ -185,6 +185,7 @@
       <DependentUpon>Settings.settings</DependentUpon>
     </Compile>
     <Compile Include="Models\TrackPersons.cs" />
+    <Compile Include="SQLStringComparison.cs" />
     <Compile Include="Synchronizers\AlbumArtSynchronizer.cs" />
     <Compile Include="Synchronizers\AlbumSynchronizer.cs" />
     <Compile Include="Synchronizers\ArtistPersonsSynchronizer.cs" />
